third_party/WebKit/LayoutTests/inspector/styles/css-live-edit-expected.txt

Tests that styles are updated when live-editing css resource.


Running: testLiveEdit
[expanded] 
display: block;
    div - block user agent stylesheet
font-size: 20px;
    body - 20px css-live-edit.css:1

[expanded] 
element.style  { ()
